Change project-detail page to route parameter variable
